requirements.txt
streamlit>=1.35
python-dotenv>=1.0

langchain==1.2.0
langchain-community==0.4.1
langchain-core==1.2.4

langchain-groq>=0.1.9
langchain-huggingface>=0.0.3
# langchain-openai>=0.1.22
langchain-chroma>=0.1.2
chromadb>=0.5.5

pypdf>=4.2.0
tiktoken>=0.7.0
sentence-transformers

ddgs
wikipedia>=1.4.0
arxiv>=2.1.0
validators>=0.28.0

youtube-transcript-api>=1.2.3
beautifulsoup4
requests

rank-bm25>=0.2.2